Adapt project to run on NB 8.2; include traineddata to .gitignore
