+++ .gitignore
@@ -1,0 +1,3 @@
+/nbproject/private/
+/build/
+/dist/
