+++ lib/nblibraries.properties
@@ -1,0 +1,12 @@
+libs.CopyLibs.classpath=\
+    ${base}/CopyLibs/org-netbeans-modules-java-j2seproject-copylibstask.jar
+libs.CopyLibs.displayName=CopyLibs Task
+libs.CopyLibs.prop-version=2.0
+libs.junit_4.classpath=\
+    ${base}/junit_4/junit-4.12.jar
+libs.junit_4.displayName=JUnit 4.12
+libs.junit_4.javadoc=\
+    ${base}/junit_4/junit-4.12-javadoc.jar
+libs.junit_4.prop-maven-dependencies=junit:junit:4.12:jar
+libs.junit_4.src=\
+    ${base}/junit_4/junit-4.12-sources.jar
