+++ nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+nbbuild.xml.data.CRC32=998548c8
+nbbuild.xml.script.CRC32=1fba10a8
+nbbuild.xml.stylesheet.CRC32=8064a381@1.80.1.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=998548c8
+nbproject/build-impl.xml.script.CRC32=3da49aa3
+nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
