--- src/JFXoverlay.fxml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.text.*?>
-<?import javafx.scene.control.CheckBox?>
-
-<BorderPane prefHeight="517.0" prefWidth="546.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userrecognition.FXController">
-   <top>
-		<VBox>
-   <children>
-   			<HBox alignment="CENTER" spacing="10">
-   				<padding>
-   					<Insets bottom="10" top="10" />
-   				</padding>
-      <children>
-      				<CheckBox fx:id="haarClassifier" onAction="#haarSelected" text="Haar Classifier" />
-      				<CheckBox fx:id="lbpClassifier" onAction="#lbpSelected" text="LBP Classifier" />
-      </children>
-   			</HBox>
-   			<Separator />
-   </children>
-		</VBox>
-	</top>
-	<center>
-		<VBox alignment="CENTER">
-			<padding>
-				<Insets left="10" right="10" />
-			</padding>
-   <children>
-   			<ImageView fx:id="originalFrame" />
-   </children>
-		</VBox>
-	</center>
-	<bottom>
-		<HBox alignment="CENTER">
-			<padding>
-				<Insets bottom="25" left="25" right="25" top="25" />
-			</padding>
-   <children>
-   			<Button fx:id="cameraButton" alignment="center" disable="true" onAction="#startCamera" text="Start camera" />
-            <CheckBox fx:id="newUser" onAction="#newUserSelected" text="New user">
-               <HBox.margin>
-                  <Insets left="40.0" />
-               </HBox.margin>
-            </CheckBox>
-            <TextField fx:id="newUserName" prefHeight="25.0" prefWidth="130.0" promptText="Enter your name">
-               <HBox.margin>
-                  <Insets left="20.0" />
-               </HBox.margin>
-            </TextField>
-            <Button fx:id="newUserNameSubmit" onAction="#newUserNameSubmitted" text="Submit">
-               <HBox.margin>
-                  <Insets left="20.0" />
-               </HBox.margin>
-            </Button>
-   </children>
-		</HBox>
-	</bottom>
-</BorderPane>
