+++ src/userrecognition/Main.java
@@ -1,0 +1,63 @@
+package userrecognition;
+
+import org.opencv.core.Core;
+
+import javafx.application.Application;
+import javafx.stage.Stage;
+import javafx.scene.Scene;
+import javafx.scene.layout.BorderPane;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+
+/**
+ * The main class for a JavaFX application. It creates and handle the main
+ * window with its resources (style, graphics, etc.).
+ * 
+ * This application handles a video stream and try to find any possible human
+ * face in a frame. It can use the Haar or the LBP classifier.
+ * 
+ * @author Luigi De Russis / Igor @ HeroinSoul / Jim O'Connorhorrill @ cuerobotics
+ * @version 1.3 (2016-08-04)
+ * @since 1.0 (2014-01-10)
+ * 
+ */
+public class Main extends Application
+{
+	@Override
+	public void start(Stage primaryStage)
+	{
+		try
+		{
+                        
+			FXMLLoader loader = new FXMLLoader(getClass().getResource("JFXoverlay.fxml"));
+			BorderPane root = (BorderPane) loader.load();
+			// set a whitesmoke background
+			root.setStyle("-fx-background-color: whitesmoke;");
+			// create and style a scene
+			Scene scene = new Scene(root, 800, 600);
+			scene.getStylesheets().add(getClass().getResource("application.css").toExternalForm());
+			// create the stage with the given title and the previously created
+			// scene
+			primaryStage.setTitle("Face Detection and Tracking");
+			primaryStage.setScene(scene);
+			// show the GUI
+			primaryStage.show();
+			
+			// init the controller
+			FXController controller = loader.getController();
+			controller.init();
+		}
+		catch (Exception e)
+		{
+			e.printStackTrace();
+		}
+	}
+	
+	public static void main(String[] args)
+	{
+		// load the native OpenCV library
+		System.loadLibrary(Core.NATIVE_LIBRARY_NAME);
+		
+		launch(args);
+	}
+}
