+++ src/userrecognition/application.css
@@ -1,0 +1,1 @@
+/* JavaFX CSS - Leave this comment until you have at least create one rule which uses -fx-Property */
